pyproject.toml
[project]
name = "fabricflow"
description = "A code-first approach for MS Fabric data pipelines."
version = "0.0.1"
dependencies = ["semantic-link-sempy>=0.8.0"]
readme = "README.md"
license = "MIT"
authors = [{ name = "Parth Lad", email = "parthlad3915@gmail.com" }]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools]
include-package-data = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.commands]
build = "python -m build"
install = "pip install -e ."
test = "pytest"
testpypi = "twine upload --repository testpypi dist/*"
pypi = "twine upload --repository pypi dist/*"
